fastapi-app/requirements.txt
fastapi
fastapi[standard]
httpx
opentelemetry-api
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
opentelemetry-sdk
uvicorn
